requirements.txt
matplotlib==3.10.0
openai==1.59.5
pandas==2.2.3
python-dotenv==1.0.1
streamlit==1.28.2
streamlit_chat==0.1.1
openpyxl==3.1.5
streamlit_feedback==0.1.4